+++ test/date.spec.ts
@@ -1,6 +1,6 @@
 import assert from 'assert';
 import { format } from 'date-fns';
-import { dayDiff, today, isRealDate } from '../dist';
+import { dayDiff, today, isValidDate } from '../dist';
 
 describe('Date', () => {
 	it('dayDiff', (done) => {
@@ -16,12 +16,12 @@ describe('Date', () => {
 		done();
 	});
 
-	it('isRealDate', (done) => {
-		assert.strictEqual(isRealDate('2021-01-01'), true);
-		assert.strictEqual(isRealDate('2021-02-28'), true);
-		assert.strictEqual(isRealDate('2021-02-29'), false);
-		assert.strictEqual(isRealDate('2021-03-32'), false);
-		assert.strictEqual(isRealDate('2021-13-01'), false);
+	it('isValidDate', (done) => {
+		assert.strictEqual(isValidDate('2021-01-01'), true);
+		assert.strictEqual(isValidDate('2021-02-28'), true);
+		assert.strictEqual(isValidDate('2021-02-29'), false);
+		assert.strictEqual(isValidDate('2021-03-32'), false);
+		assert.strictEqual(isValidDate('2021-13-01'), false);
 		done();
 	});
 });